% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaoverfit.R
\name{print.sample_size_rec}
\alias{print.sample_size_rec}
\title{Print Method for sample_size_rec Objects}
\usage{
\method{print}{sample_size_rec}(x, ...)
}
\arguments{
\item{x}{A sample_size_rec object.}

\item{...}{Additional arguments (unused).}
}
\value{
No return value, called for side effects (printing to console).
}
\description{
Print Method for sample_size_rec Objects
}
