% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaoverfit.R
\docType{package}
\name{metaoverfit-package}
\alias{metaoverfit}
\alias{metaoverfit-package}
\title{metaoverfit: Optimism-Corrected Heterogeneity in Meta-Regression}
\description{
Tools for detecting and correcting overfitting in meta-regression analyses
through cross-validation and bootstrap methods.
}
\author{
\strong{Maintainer}: Mahmood Ahmad \email{mahmood.ahmad2@nhs.net}

}
\keyword{internal}
