% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaoverfit.R
\name{print.metaoverfit}
\alias{print.metaoverfit}
\title{Print Method for metaoverfit Objects}
\usage{
\method{print}{metaoverfit}(x, ...)
}
\arguments{
\item{x}{metaoverfit object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for metaoverfit Objects
}
